requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
numpy>=2.0.0
pandas>=2.2.2
scikit-learn>=1.5.0
tensorflow-cpu>=2.17.0
psycopg2-binary==2.9.9
stable-baselines3
shimmy